Add basic (non)contiguous subarray support
The compile time constant MPI_SUBARRAYS_SUPPORTED is changed to true now.

To test if a subarray is contiguous, one may use CFI_is_contiguous(cdesc). But note that
CFI_is_contiguous(cdesc) is only applicable to arrays. When cdesc is a descripitor for a scalar,
CFI_is_contiguous(cdesc) returns false. But apparently, scalars are contiguous in MPI's viewpoint.
So we add a check against scalars.

No review since F08 binding is experimental now.