pamid: force context post on when using 'per-object' locks
The previous code would attempt to run in 'latency optimization mode' if
the application requested !MPI_THREAD_MULTIPLE and yet still was linking
with a 'per-object' mpich library - which is optimized for throughput.
This means that in !MPI_THREAD_MULTIPLE:
- context post was disabled
- async progress was disabled
- using multiple contexts was disabled

This attempt to give users "better" performance in a fundamentally
flawed configuration would cause the pamid adi to hang on acquiring a
context lock. For example, consider:
1. work function is posted to a context
2. thread acquires the context lock
3. thread advances the context
4. work function invoked, then attempts to initiate communication
5. WITH CONTEXT POST OFF the thread will attempt to acquire the context
   lock before it directly invokes the communcation function
6. HANG

A complete solution would be to identify all code paths that might
result in this situation and instead avoid acquiring a lock that is
already held.

It should be noted that this is a bug in the pamid adi and NOT in the
PAMI library iteself. The PAMI interface is explicit in that the
PAMI_Context_lock() is non-recursive - for performance reasons.