pamid: fix request completion race condition
In a multi-threaded environment where one thread is waiting for a
request to complete, the instant the request completion counter is set
to zero by a different thread the waiting thread will begin to free the
request resources. The completing thread must not touch any memory from
the old request object after setting the completion count to zero.

See ticket #2103