use --enable-new-dtags in compiler scripts
The use of LD_LIBRARY_PATH is not sufficient to override RPATH tags
encoded in an ELF binary. This hinders users looking to swap ABI
compatible versions of MPI at runtime. To enable this use, the compiler
scripts should use --enable-new-dtags, which sets RUNPATH instead of
RPATH in the binary, and will be checked after LD_LIBRARY_PATH. Thanks
to Intel for suggesting this change.

Signed-off-by: Pavan Balaji <balaji@anl.gov>