Fixes building MPI_Comm_idup for single-thread case
The patch 05eeccb5 causes failure in building with
'--enable-threads=single'.  It fixes the MPI_Comm_idup multi-threaded
case but messes up with the single-threaded case.  This commit reverts
back to the old code before 05eeccb5 for the single-threaded case and
keeps the fix for multi-threaded case.

Note that the old code is still not correct for multiple communicators.
See #1935.

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>