Add CVAR (# of completed reqs) to control when to stop in RMA GC function
Add a CVAR to let the user specify the threshold for number of
completed requests the runtime finds before it stops trying to
find more completed requests in garbage collection function. It
may make the runtime to find more completed requests, but may also
cause significant overhead due to visiting too many requests.

Note that the default value is set to 1 in order to be in
accordance with the performance of original implementation.

Note that in garbage collection function, if runtime finds a chain
of completed RMA requests, it will temporarily ignore this CVAR
and try to find continuous completed requests as many as possible,
until it meets an incomplete request.

Signed-off-by: Pavan Balaji <balaji@anl.gov>