pamid: create memregions at an offset of the shared memory allocation
For MPI_Win_allocate_shared() the memory regions that are exchanged all
pointed to the begining of the shared memory allocation instead of at
the portion of the shared memory allocated to each process. When
MPI_Put(), for example, issues a PAMI_Rput() it uses an offset of zero
to specify the begining of the peer's window. As the exchanged memory
regions for each peer rank were not created at the start of each peer
rank's allocation, data corruption and other bad things occur.