pamid: remove blocking shmem mutex; remove shmem CAS/FOP optimizations
The blocking pthread mutex in the shared memory window causes a deadlock
on bgq - perhaps because the messaging state is not advanced while the
thread is waiting for the mutex to be released.

Removing this blocking mutex resolves the bgq failures:

 - rma/strided_putget_indexed_shared
 - rma/strided_getacc_indexed_shared

With the calls to the blocking mutex removed the CAS and FOP functions
are not atomic. Solution is to remove the shared memory optimization and
instead use the common (network) code path.

Removing these shared memory optimizations from CAS/FOP resolves the
bgq hangs:

 - rma/mutex_bench_shared