Add a CVAR to turn off collective alias checking
Add a new CVAR (MPIR_CVAR_COLL_ALIAS_CHECK) that allows the user to turn off
alias checking in collective operations. In 5b4bdef9, alias checking was
expanded to be more correct according to the standard, but will cause some
previously misbehaving applications to fail even when they might not have
displayed that behavior before. This patch lets users turn off checking if
they accept the risk and don't want this specific check to be done. All error
checking can still be turned off at configure time as usual. By default, this
new CVAR will be set to `1` to indicate that checking is on.

Fixes #2117

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>