Clean up MPI_Comm_idup code for '--enable-threads=single' case.
This patch clean the code for the MPI_Comm_idup single-threaded case to
use the same code of the multi-threaded case. In order to do this, three
variables are made visible to both single-threaded and multi-threaded
case: mask_in_use, lowestContextId and lowestTag.

Also, before this patch, #1935 has been fixed for multi-threaded case
but will still fail for single-threaded case. After this patch, the
single-threaded case will be fixed.

Here is a history of related changes: The patch [05eeccb5] fixed
MPI_Comm_idup for multi-threaded case, but forgot to fix the
single-threaded case. A temporary patch [18de1dbe] is used to revert
the single-threaded case to use old code before multi-thread patch.
The current patch fixes the second patch.

See also #1935.

Signed-off-by: Pavan Balaji <balaji@anl.gov>