Lustre used constants w/o header file
If we are going to use UINT_MAX and USHRT_MAX, something needs to
bring in limits.h.  Martin Audet <Martin.Audet@nrc-cnrc.gc.ca> reported
a build failure because we did not include limits.h.  In addition to his
one-line patch I threw in a configure-time check just in case.