Add CVAR (# of tested reqs) to control when to stop in RMA GC function
When cleanning up completed requests, the original RMA implementation
keeps traversing the op list until it finds a completed request. This
may cause significant O(N) overhead if there is no completed request
in the list. We add a CVAR to let the user control the number of visited
requests as a fixed value.

Note that the default value is set to (-1) in order to be in accordance
with the performance of orignal implementation.

Note that in garbage collection function, if runtime finds a chain
of completed RMA requests, it will temporarily ignore this CVAR
and try to find continuous completed requests as many as possible,
until it meets an incomplete request.

Signed-off-by: Pavan Balaji <balaji@anl.gov>