Fix bug in PMI keyval caching logic.
When we were flushing the PMI keyval cache, the cache was flushed in
the form of a single very long PMI command.  The server would
interpret this as a single PMI command with many arguments.  However,
if the cache is larger than the allowed number of PMI arguments, we
might run out of the allocated space.  This patch connects these two
values, so the cache can never overrun the number of PMI arguments.

Thanks to Norio Yamaguchi @ Riken for reporting this issue.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>