Fix bug: add barrier semantic in FENCE for SHM ops.
When SHM is allocated for RMA window, operations are completed
eagerly (as soon as they are posted by the user), therefore we
need barrier semantics in the FENCE that opens an epoch to prevent
SHM ops happening on target process before that target process
starts an epoch.

Note that we need memory barrier before and after synchronization
calls in both FENCEs that starts and ends an epoch to guarantee the
ordering of load/store operations with synchronizations.

See #2041.

Signed-off-by: Pavan Balaji <balaji@anl.gov>