Add warning about memory tracing overhead.
When the debugging option about memory usage tracing is turned on,
additional memory overhead for tracing may be significant especially
when a large number of small memory allocations are requested, as
reported in #2068. However, since the tracing data is necessary to
keep track of memory usage, it cannot be removed. Here, we provide
a warning message when the memory overhead for tracing is over the
threshold value.

See #2068

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>