deal better with large user-defined types
this optimization in ROMIO's noncontigous type handling detects if a
user-defined type will fit into a single write request (e.g. a call to
MPI_Type_create_subarray that was actually the entire array).  Code was
passing in a count of bytes, which can overflow an int. since we know
the type is contiguous, simply pass a count of that type (not bytes) to
write_contig

Signed-off-by: Mohamad Chaarawi <chaarawi@hdfgroup.org>