pamid: Data integrity error hit in MPI_Win_allocate_shared windows(LE)
MPI_Win_shared_query() currently returns the base address of the passed in rank.
All tasks other than task 0 assume that the first set of data belongs to task 0,
second set belongs to task 1 and so on. The assumption caused the data integrity
problem found by win_shared.c.

To fix the problem, instead of returning the base address of a rank,
MPI_Win_shared_query() should return the base address of shared segment
allocated for the window.

(ibm) D198663

Signed-off-by: Michael Blocksome <blocksom@us.ibm.com>