Add some basic resilience to allred_group
If a process is dead, collectives still do all of the communictaions to
prevent a deadlock. However, if we just skip the part where the data is
updated in the allreduce_group function, we can let it be slightly more
resilient to failures and possibly even produce a correct answer in the
presence of a failure.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>