Don't start enums with 0.
This is to help with debugging.  Zero is too common a value, and is
often set automatically by the system if not initialized.  Starting at
a different value helps us catch uninitialized cases more easily.

We pick "42" as our magic number as it is the answer to the ultimate
question of life, the Universe, and everything.

Signed-off-by: Wesley Bland <wbland@anl.gov>