Allow channel packet handlers in the CH3 structure.
In nemesis we were managing a separate enum for the nemesis packet
handlers, which was somehow being force compared against the CH3
packet handler enums.  This is weird and a recipe for bugs.  This
patch unifies them into a single enum, but allowing the channel to
declare its own packet handlers into the structure.

Signed-off-by: Wesley Bland <wbland@anl.gov>