Destroy request object before setting it to NULL.
When an isendv fails in MPIDI_CH3_EagerSyncNoncontigSend, the request is set
to NULL before returning it back to the caller. Unfortunately, the request is
not allocated inside this function so if we pass back NULL, we lose the handle
on the request. If we're going to return NULL, we need to make sure the
request is destroyed before giving it back.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>