Reorganize mpi.h for MPIX err codes
The mpi.h header already had one MPIX error code in the middle of the regular
MPI error codes. This is bad for ABI compatibility since if that error code
ever needs to change, it might cause problems. To avoid this, we'll now have a
new value called MPI_ERR_FIRST_MPIX which is bigger than MPICH_ERR_LAST_CLASS.
All MPIX error classes will be based on this value. If anything above that
value ever changes, it's not a problem because it's MPIX and not part of the
ABI agreement.

There is a gap between MPICH_ERR_LAST_CLASS and MPICH_ERR_FIRST_MPIX because
sock is currently using these values for some of its internal error codes.
Someday in the future, we could consider removing this gap if sock goes away.

This commit also does some minor reordering of the error codes within the file
(not their values) for readability reasons.

Signed-off-by: Pavan Balaji <balaji@anl.gov>