Fixed compile warnings in MPI_File_f2c with --disable-romio
MPI_File_f2c() is expected to return a value of type MPI_File. However,
the old code returns MPI_File*. We do a type casting here. Whether the
returned value is correct is not important since romio is disabled and we
just want to make the compiler happy.

Fixes #2140

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>