Introduce MPICH_ERR_LAST_MPIX
There were a few places in MPICH where the error class was being checked
against MPICH_ERR_LAST_CLASS and being flagged as invalid if it was too large.
This is incorrect now that we have a new space for MPIX error codes. Add
MPICH_ERR_LAST_MPIX as a way of keeping track of what the actual last valid
error class is.

Signed-off-by: Junchao Zhang <jczhang@mcs.anl.gov>