Correctly report the error class in receive queue
The receive queue had some hacky ways of reporting errors related to process
failure that didn't really match up with the way the codes should be returned
correctly. This patch sets the correct error class in the correct place and
doesn't require extra logic in dequeue_and_set_error to set the class itself.

This seems to get a couple of the tests to pass in non-debug mode.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>