Improves random number generation in shfp_fname.c
The original code uses MPI_Wtime to generate a mostly random double
number, and convert it to int. The integer is then used as a part of a
shared file name. Replaced with rand().

It fixes the warning messages of 'Warning: ieee_inexact is signaling'
signaled by PG Fortran compiler.

Fixes #2071

Signed-off-by: Rob Latham <robl@mcs.anl.gov>