Fix error case for MPIDI_Request_create_null_rreq
For some reason, the error case code between MPIDI_Request_create_rreq and
MPIDI_Request_create_null_rreq was different. This is odd, because both macros
take FAIL_ as an argument which is executed directly in the error case of
create_rreq, but not in null_req. This commit makes the two act the same and
updates the only two calls to the function that existed in the code.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>