Don't put revoked requests into the receive queue
When a message is received, if the communicator has already been revoked, we
shouldn't bother keeping the message since it's now invalid (unless its for an
AGREE or SHRINK request). Instead, just drop the request and return a null
request to signal the calling function that the request was ignored.

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>