Added a test for flush on shared window.
Flush should guarantee operations are finished on both origin and target
side. However, flush may return before the completion on target side in
MPI implementation. It makes an error in this case: P0 and P1 allocate a
shared window, and P2 locks both of them; P2 first put to P0 and flush,
then get the updated data from P1. The put may complete on P0 after the
completion of get on P1.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>
Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>