Enabled SHM segments detection in MPI_Win_create
First, cache every SHM window created by Win_allocate or
Win_allocate_shared into a global list, and unlink it in Win_free.

Then, when user calls Win_create for a new window, check user specified
buffer and comm. Enable local SHM communicaiton in the new window if it
matches a cached SHM window. It is noted that all the shared resources
are still freed by the original SHM window.

Matching a SHM window must satisfy following two conditions:
1. The new node comm is equal to, or a subset of the SHM node comm.
(Note that in the other cases where two node comms are overlapped,
although the overlapped processes could be logically shared, it is not
supported for now. To support this, we need to fist modify the implementation
of RMA operations in order to remember shared status per target but not
just compare its node_id).
2. The buffer is in the range of the SHM segment across local processes
in original SHM window (a contigunous segment is mapped across local
processes regardless of whether alloc_shared_noncontig is set).

Resolves #2161

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>