mpi/comm: Fix MPI_Intercomm_merge
Previous code is based on a trick to get true context id using
invalid communicator. There is a call of MPIR_Get_contextid()
for new communicator that is not completelly built (no comm_commit
and comm_hook calls).
netmod/mxm uses comm_hook and can not work with this trick.
This changes allow to avoid call of invalid communicator using
temporary (intermediate) communicator.

Signed-off-by: Igor Ivanov <Igor.Ivanov@itseez.com>
Signed-off-by: Pavan Balaji <balaji@anl.gov>