Bug-fix: avoid wrong synchronization in MPI_Win_complete.
recv_post_msgs function on origin process is used to wait
for synchronization from target processes, and it can be
called from MPI_Win_start (for targets on SHM) and
MPI_Win_complete (for targets not on SHM). Here we fix the
bug so that origin will not wait for SHM targets again in
MPI_Win_complete.

Signed-off-by: Pavan Balaji <balaji@anl.gov>