Remove Blue Gene/Q specific MPID_VCR_GET_LPIDS
Since Blue Gene/Q does not support dynamic tasking there was only 1
element in the MPID_VCR_t data structure so a shortcut was taken
to avoid a malloc and free of a new list of pami_task_t in a form
the pami geometry creation was expecting.  However it seems an
array of structures with 1 pami_task_t element in it is not exactly the same
in memory as an array of pami_task_t  themselves so the pami
geometry creation was failing.  The fix is to simply do what
all other platforms do and malloc a separate list of pami_task_t.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>