MPIDI_Coll_comm_create unsetting default collops for inTERcomms
There are no pami optimized collectives available for inTERcomms
so when MPIDI_Coll_comm_create is called it checks to see
if the comm is an inTRAcomm and if not just returns,
however it was doing the new malloc for comm->coll_fns
before the if-check which was unsetting the MPICH defaults.
The solution is to move the malloc after the inTRAcomm
if-check.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>