MPICH-specific initialization of hcoll.
In some cases, we cannot let hcoll use whatever transport it needs.
For example, ch3:sock assumes that while blocking the next event will
come over the socket channel.  If HCOLL decides to use a different
transport (such as mxm) and the next event comes on that transport,
this can result in a deadlock.  In this patch, we let the channel
specify what transports it can accept.

Signed-off-by: Devendar Bureddy <devendar@mellanox.com>