mpi/coll: Fix incorrect parameter check
Fixed wrong parameter check condition for MPI_Iallgather and MPI_Iallgatherv
-1 is valid value for sendcount in case MPI_IN_PLACE
MPI spec says:
The in place option for intracommunicators is specified by passing the value
MPI_IN_PLACE to the argument sendbuf at all processes. In such a case, sendcount and
sendtype are ignored, and the input data of each process is assumed to be in the area where
that process would receive its own contribution to the receive buffer.

Signed-off-by: Igor Ivanov <Igor.Ivanov@itseez.com>
Signed-off-by: Pavan Balaji <balaji@anl.gov>