portals4: remove pointer update from unmatched improbe
Previously, the message pointer in an improbe call in the portals4
netmod layer was set to MPI_MESSAGE_NULL if there was no match. This
is incorrect because the ch3 layer eventually returns either a valid
pointer or NULL. As ch3 already starts the value at NULL, we can just
omit the update.

Signed-off-by: Pavan Balaji <balaji@anl.gov>