Bug-fix: trigger final req handler for receiving derived datatype.
There are two request handlers used when receiving data:
(1) OnDataAvail, which is triggered when data is arrived;
(2) OnFinal, which is triggered when receiving data is finished;

When receiving large derived datatype, the receiving iov can be divided
into multiple iovs. The OnDataAvail handler is set to iov load function
when still waiting for remaining data. However, such handler should be
set to OnFinal when starting receiving the last iov.

The original code does not set OnDataAvail handler to OnFinal at end.
This patch fixes this bug.

Note that this bug only appears in RMA calls, because only the RMA
packet handers need to specify OnFinal.

Resolve #2189.

Signed-off-by: Xin Zhao <xinzhao3@illinois.edu>