MP_IOTASKLIST error checking
PE users may manually specify the MP_IOTASKLIST for explicit aggregator
selection.  Code needed to be added to verify that the user
specification of aggregators were all valid.

Do our best to maintain the old PE behavior of using as much of the
correctly specified MP_IOTASKLIST as possible and issuing what it
labeled error messages but were really warnings about the incorrect
portions and functionally just ignoring it, unless none of it was usable
in which case it fell back on the default.

Signed-off-by: Rob Latham <robl@mcs.anl.gov>