portals4: set reasonable interface limits
Set reasonable limits for maximum unexpected headers and EQs at init
time. We accomplish this with a pre-init stage where we fill in a limits
struct with the system defaults, increase certain values (if they are not
set already in the environment), then do the real init.

If the "desired" limits structure had a way to allow default values for
limits we don't care about, the pre-init stage could go away.

Signed-off-by: Antonio J. Pena <apenya@mcs.anl.gov>