Add a queue for shm RTS messages
RTS messages (the first part of the LMT sequence) had no way of being
cancelled if an error occurred. This adds a small queue that keeps track
of these messages. If a failure is detected, the message is removed from
the queue and the associated request is cancelled to get out of the
progress engine.

See #1945

Signed-off-by: Huiwei Lu <huiweilu@mcs.anl.gov>